refactor and add support for rebasing image
