+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/cli/menu/ImageBuildOptions.java
@@ -1,0 +1,38 @@
+// Copyright (c) 2019, Oracle Corporation and/or its affiliates.  All rights reserved.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package com.oracle.weblogic.imagetool.cli.menu;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import picocli.CommandLine.Option;
+import picocli.CommandLine.Unmatched;
+
+public class ImageBuildOptions extends CommonOptions {
+    @Option(
+        names = {"--latestPSU"},
+        description = "Whether to apply patches from latest PSU."
+    )
+    boolean latestPSU = false;
+
+    @Option(
+        names = {"--patches"},
+        paramLabel = "patchId",
+        split = ",",
+        description = "Comma separated patch Ids. Ex: 12345678,87654321"
+    )
+    List<String> patches = new ArrayList<>();
+
+
+    @Option(
+        names = {"--opatchBugNumber"},
+        description = "the patch number for OPatch (patching OPatch)"
+    )
+    String opatchBugNumber = "28186730";
+
+
+    @Unmatched
+    List<String> unmatchedOptions;
+
+}
