+++ imagetool/src/main/resources/probe-env/test-update-env.sh
@@ -12,6 +12,7 @@ fi
 
 if [[ ! -z "$JAVA_HOME" ]]; then
   echo JAVA_HOME="$JAVA_HOME"
+  echo JAVA_PATH="$(readlink -f $JAVA_HOME)"
 fi
 
 if [[ ! -z "$ORACLE_HOME" ]]; then
@@ -32,4 +33,8 @@ if [[ ! -z "$ORACLE_HOME" ]]; then
     #echo $LSINV_TEXT | base64 > /tmp/b64
     echo LSINV_TEXT=$(base64 /tmp/lsout)
   fi
+fi
+
+if [[ ! -z "$DOMAIN_HOME" ]]; then
+  echo DOMAIN_HOME="$DOMAIN_HOME"
 fi
